Rename create with StoreViewComponent
